pyproject.toml
[tool.poetry]
name = "aiaccel2gpt-v5"
version = "0.1.0"
description = ""
authors = ["Takashi OKADA <tk.okada.github@gmail.com>"]
readme = "README.md"
packages = [{include = "aiaccel2gpt_v5"}]

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0.15"
pydantic = "^1.10.7"
fastapi = "^0.95.2"
click = "^8.1.3"
uvicorn = "^0.22.0"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.3.0"
sphinx = "^7.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
